Change Damage nbt to i32 instead of i16
In NBTExplorer on a vanilla world, I see that damage is an int, not a short.

+++ core/anvil/src/entity.rs
@@ -230,7 +230,7 @@ where
 #[derive(Default, Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct ItemNbt {
     #[serde(rename = "Damage")]
-    pub damage: Option<i16>,
+    pub damage: Option<i32>,
     // TODO enchantments, display name, ...
 }
